requirements.txt
streamlit
openai
llama-cpp-python
pytesseract
Pillow
